Move test_inj_calcs to test_fuel_controller
