+++ test/test_schedule_calcs/main.cpp
@@ -5,11 +5,9 @@
 void runAllScheduleCalcTests(void)
 {
   extern void test_calc_ign_timeout();
-  extern void test_calc_inj_timeout();
   extern void test_adjust_crank_angle();
 
   test_calc_ign_timeout();
-  test_calc_inj_timeout();
   test_adjust_crank_angle();
 }
