--- test/test_schedule_calcs/test_calcs_common.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <stdint.h>
-
-void setEngineSpeed(uint16_t rpm, int16_t max_ign);
